Moved LocalTlsSetup from the http module
